+++ xpdacq/tests/test_run_calib.py
@@ -1,8 +1,7 @@
 import os
+import pytest
 import shutil
 import uuid
-
-import pytest
 from pkg_resources import resource_filename as rs_fn
 from pyFAI.calibrant import Calibrant, ALL_CALIBRANTS
 
@@ -53,6 +52,7 @@ def test_configure_sample_info_md(sample_name, phase_info, tag, sample_md):
     assert parsed_sample_md == sample_md
 
 
+@pytest.mark.skip
 def test_calib_md(fresh_xrun, exp_hash_uid, glbl, db):
     xrun = fresh_xrun
     # calib run