+++ news/news.rst
@@ -1,12 +1,12 @@
 **Added:**
 
-* Add `xpdacq.factory.BasicPlans` for the multi-detector plan for both XRD and PDF data collection in one run.
+* Add `xpdacq.plans.XrayBasicPlans` for the multi-detector plan for both XRD and PDF data collection in one run.
 
-* Add `xpdacq.factory.MultiDistPlans` for the moving detector measurement for both XRD and PDF data collection in one run.
+* Add `xpdacq.plans.MultiDistPlans` for the moving detector measurement for both XRD and PDF data collection in one run.
 
 * Add `xpdacq.devices.CalibrationData`, a class to store the calibration data of a detector in configuration attributes.
 
-
+* Add `xpdacq.plans.XrayBasicPlans.create_dark_frame_processor` for the new dark protocol that puts dark in a side stream.
 
 **Changed:**